cc_jackfruit_final/student_welfare_service/requirements.txt
Flask
Flask-MySQLdb
mysqlclient
Flask-WTF
python-dotenv